шедуллер перенесен в контейнер, запускается при старте application
